Added support for conditions precedence in table join conditions. Closes #18.
SQLSelectTableJoinConditions objects can now be created and a collection of conditions can be added to an SQLSelectTableJoinConditions object to group a set of conditions and to ensure the sub-conditions are wrapped in parentheses and executed logically first as per the mathematical order of operations.

Removed the reference to the Friend SQLSelectTableJoinConditions.Parent property because this is not utilised by the library.

SQLSelectTableJoinConditions.Add(String, ComparisonOperator, String) is only accessible if used when the parent SQLSelectTableJoin is used and it remains only for backward compatibility reasons. Created issue #20 to indicate that this method should be marked as obsolete.