Corrected multiple users locking the same object. Closes #34.
The locking is achieved by attempt to insert into the lock table which contains a unique index. It the insert fails then the record is already locked because the insert breaks the database table's key constraint. However, the UserID was incorrectly included in the unique index / key which meant that attempts by multiple users to insert into the table would not fail.