Added RootContainerObject concept. Closes #8.
This allows a root container object to be accessed from any object in the model. The overrides the existing functionality of providing a database for a root collection. If there are multiple levels to the hierarchy then changing the root collection object to utilise a RootContainer will allow all children to automatically find the reference to the parent root container object.