Added support SQLExpression operator overloading. Closes #52.
Provides the powerful ability to generate SQL expressions for arithmetic operations and string concatenation using regular .NET code. Any arithmetic operator or string concatenation can be applied to an SQLExpression and the appropriate operator is translated into the appropriate SQL command. This significantly simplifies the .NET code and also simplifies the process of generating the SQL code, because normal .NET operators can be utilised.

For example: "New SQLFieldExpression("Field") Mod 10" translates to SQL "([Field] % 10)". The expression can be used in an SQLSelect, SQLInsert, SQLDelete or SQLUpdate class (amongst others).